move data browser source files into applab/dataBrowser
