+++ apps/src/applab/dataBrowser/DataProperties.jsx
@@ -3,11 +3,11 @@
  */
 
 import AddKeyRow from './AddKeyRow';
-import { DataView } from './constants';
+import { DataView } from '../constants';
 import EditKeyRow from './EditKeyRow';
 import Radium from 'radium';
 import React from 'react';
-import { changeView } from './redux/data';
+import { changeView } from '../redux/data';
 import { connect } from 'react-redux';
 import * as dataStyles from './dataStyles';
